home2

+++ ft_bzero.c
@@ -3,17 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_bzero.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: antgarci <antgarci@student.42malaga.c      +#+  +:+       +#+        */
+/*   By: codespace <codespace@student.42.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/04/29 18:55:02 by antgarci          #+#    #+#             */
-/*   Updated: 2026/04/29 18:55:20 by antgarci         ###   ########.fr       */
+/*   Updated: 2026/04/29 21:42:11 by codespace        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
+
 void	ft_bzero(void *str, size_t n)
 {
-	int	i;
+	size_t			i;
 	unsigned char	*ptr;
 
 	ptr = (unsigned char *)str;
